moved constants.py back to the top level
